fix crashes on kasan build
why: Read size 2 at addr ffffffc00de77490 by task init
how: change from u16 to u8
Reviewed-on: https://gerrit.mot.com/3029698
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Submit-Approved: Jira Key

+++ drivers/usb/typec/sc2150/type_sc2150.c
@@ -135,7 +135,7 @@ static int sc2150_read_device(void *client, u32 reg, int len, void *dst)
 		ret = i2c_smbus_read_i2c_block_data(i2c, reg, len, dst);
 		t2 = local_clock();
 		SC2150_INFO("%s del = %lluus, reg = 0x%02X, len = %d, val = 0x%08X\n",
-			    __func__, (t2 - t1) / NSEC_PER_USEC, reg, len, *(u16 *)dst);
+			    __func__, (t2 - t1) / NSEC_PER_USEC, reg, len, *(u8 *)dst);
 		if (ret < 0 && count > 1)
 			count--;
 		else
